+++ sdk/eventhubs/azure-messaging-eventhubs/src/test/java/com/azure/messaging/eventhubs/EventHubConsumerAsyncClientTest.java
@@ -68,12 +68,14 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static com.azure.core.amqp.AmqpMessageConstant.ENQUEUED_TIME_UTC_ANNOTATION_NAME;
-import static com.azure.core.util.tracing.Tracer.ENTITY_PATH_KEY;
-import static com.azure.core.util.tracing.Tracer.HOST_NAME_KEY;
 import static com.azure.core.util.tracing.Tracer.PARENT_TRACE_CONTEXT_KEY;
 import static com.azure.messaging.eventhubs.EventHubClientBuilder.DEFAULT_PREFETCH_COUNT;
+import static com.azure.messaging.eventhubs.TestUtils.assertAllAttributes;
 import static com.azure.messaging.eventhubs.TestUtils.getMessage;
+import static com.azure.messaging.eventhubs.TestUtils.getSpanName;
 import static com.azure.messaging.eventhubs.TestUtils.isMatchingEvent;
+import static com.azure.messaging.eventhubs.implementation.instrumentation.OperationName.GET_EVENT_HUB_PROPERTIES;
+import static com.azure.messaging.eventhubs.implementation.instrumentation.OperationName.GET_PARTITION_PROPERTIES;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -756,7 +758,7 @@ class EventHubConsumerAsyncClientTest {
                 List<TestMeasurement<Double>> measurements = consumerLag.getMeasurements();
                 TestMeasurement<Double> last = measurements.get(measurements.size() - 1);
                 assertEquals(Duration.between(enqueuedTime, afterReceived).toMillis() / 1000d, last.getValue(), 1);
-                assertAttributes(EVENT_HUB_NAME, e.getPartitionContext().getPartitionId(), last.getAttributes());
+                assertAllAttributes(HOSTNAME, EVENT_HUB_NAME, e.getPartitionContext().getPartitionId(), CONSUMER_GROUP, null, last.getAttributes());
                 return true;
             })
             .expectComplete()
@@ -793,7 +795,7 @@ class EventHubConsumerAsyncClientTest {
                 List<TestMeasurement<Double>> measurements = consumerLag.getMeasurements();
                 TestMeasurement<Double> last = measurements.get(measurements.size() - 1);
                 assertEquals(0, last.getValue());
-                assertAttributes(EVENT_HUB_NAME, e.getPartitionContext().getPartitionId(), last.getAttributes());
+                assertAllAttributes(HOSTNAME, EVENT_HUB_NAME, e.getPartitionContext().getPartitionId(), CONSUMER_GROUP, null, last.getAttributes());
             })
             .expectComplete()
             .verify(DEFAULT_TIMEOUT);
@@ -853,9 +855,9 @@ class EventHubConsumerAsyncClientTest {
     @Test
     void startSpanForGetProperties() {
         // Arrange
-        final Tracer tracer1 = mock(Tracer.class);
-        when(tracer1.isEnabled()).thenReturn(true);
-        EventHubsConsumerInstrumentation instrumentation = new EventHubsConsumerInstrumentation(tracer1, null,
+        final Tracer tracer = mock(Tracer.class);
+        when(tracer.isEnabled()).thenReturn(true);
+        EventHubsConsumerInstrumentation instrumentation = new EventHubsConsumerInstrumentation(tracer, null,
             HOSTNAME, EVENT_HUB_NAME, CONSUMER_GROUP, false);
         EventHubConsumerAsyncClient consumer = new EventHubConsumerAsyncClient(HOSTNAME, EVENT_HUB_NAME,
             connectionProcessor, messageSerializer, CONSUMER_GROUP, PREFETCH, false, onClientClosed, CLIENT_IDENTIFIER,
@@ -869,16 +871,23 @@ class EventHubConsumerAsyncClientTest {
         when(managementNode.getEventHubProperties()).thenReturn(Mono.just(ehProperties));
         when(managementNode.getPartitionProperties(anyString())).thenReturn(Mono.just(partitionProperties));
 
-        when(tracer1.start(eq("EventHubs.getPartitionProperties"), any(StartSpanOptions.class), any(Context.class))).thenAnswer(
+        final String expectedPartitionSpanName = getSpanName(GET_PARTITION_PROPERTIES, EVENT_HUB_NAME);
+        when(tracer.start(eq(expectedPartitionSpanName), any(StartSpanOptions.class), any(Context.class))).thenAnswer(
             invocation -> {
-                assertStartOptions(invocation.getArgument(1, StartSpanOptions.class));
+                StartSpanOptions startOpts = invocation.getArgument(1, StartSpanOptions.class);
+                assertEquals(SpanKind.CLIENT, startOpts.getSpanKind());
+                assertAllAttributes(HOSTNAME, EVENT_HUB_NAME, PARTITION_ID, CONSUMER_GROUP, null, startOpts.getAttributes());
                 return invocation.getArgument(2, Context.class)
                         .addData(PARENT_TRACE_CONTEXT_KEY, "getPartitionProperties");
             }
         );
-        when(tracer1.start(eq("EventHubs.getEventHubProperties"), any(StartSpanOptions.class), any(Context.class))).thenAnswer(
+
+        final String expectedHubSpanName = getSpanName(GET_EVENT_HUB_PROPERTIES, EVENT_HUB_NAME);
+        when(tracer.start(eq(expectedHubSpanName), any(StartSpanOptions.class), any(Context.class))).thenAnswer(
             invocation -> {
-                assertStartOptions(invocation.getArgument(1, StartSpanOptions.class));
+                StartSpanOptions startOpts = invocation.getArgument(1, StartSpanOptions.class);
+                assertEquals(SpanKind.CLIENT, startOpts.getSpanKind());
+                assertAllAttributes(HOSTNAME, EVENT_HUB_NAME, null, CONSUMER_GROUP, null, startOpts.getAttributes());
                 return invocation.getArgument(2, Context.class)
                     .addData(PARENT_TRACE_CONTEXT_KEY, "getEventHubsProperties");
             }
@@ -890,17 +899,17 @@ class EventHubConsumerAsyncClientTest {
             .expectComplete()
             .verify(DEFAULT_TIMEOUT);
 
-        StepVerifier.create(consumer.getPartitionProperties("0"))
+        StepVerifier.create(consumer.getPartitionProperties(PARTITION_ID))
             .consumeNextWith(p -> assertSame(partitionProperties, p))
             .expectComplete()
             .verify(DEFAULT_TIMEOUT);
 
         //Assert
-        verify(tracer1, times(1))
-            .start(eq("EventHubs.getPartitionProperties"), any(StartSpanOptions.class), any(Context.class));
-        verify(tracer1, times(1))
-            .start(eq("EventHubs.getEventHubProperties"), any(StartSpanOptions.class), any(Context.class));
-        verify(tracer1, times(2)).end(isNull(), isNull(), any());
+        verify(tracer, times(1))
+            .start(eq(expectedPartitionSpanName), any(StartSpanOptions.class), any(Context.class));
+        verify(tracer, times(1))
+            .start(eq(expectedHubSpanName), any(StartSpanOptions.class), any(Context.class));
+        verify(tracer, times(2)).end(isNull(), isNull(), any());
 
         verifyNoInteractions(onClientClosed);
     }
@@ -953,12 +962,6 @@ class EventHubConsumerAsyncClientTest {
         assertEquals(3, tryCount.get());
     }
 
-    private void assertStartOptions(StartSpanOptions startOpts) {
-        assertEquals(SpanKind.CLIENT, startOpts.getSpanKind());
-        assertEquals(EVENT_HUB_NAME, startOpts.getAttributes().get(ENTITY_PATH_KEY));
-        assertEquals(HOSTNAME, startOpts.getAttributes().get(HOST_NAME_KEY));
-    }
-
     private void assertPartition(String partitionId, PartitionEvent event) {
         System.out.println("Event received: " + event.getPartitionContext().getPartitionId());
         final Object value = event.getData().getProperties().get(PARTITION_ID_HEADER);
@@ -986,14 +989,5 @@ class EventHubConsumerAsyncClientTest {
             message.getMessageAnnotations().getValue().put(Symbol.valueOf(ENQUEUED_TIME_UTC_ANNOTATION_NAME.getValue()), enqueueTime);
             testPublisher.next(message);
         }
-    }
-
-
-    private void assertAttributes(String entityName, String entityPath, Map<String, Object> attributes) {
-        assertEquals(4, attributes.size());
-        assertEquals(HOSTNAME, attributes.get("hostName"));
-        assertEquals(entityName, attributes.get("entityName"));
-        assertEquals(CONSUMER_GROUP, attributes.get("consumerGroup"));
-        assertEquals(entityPath, attributes.get("partitionId"));
     }
 }