+++ sdk/eventhubs/azure-messaging-eventhubs/src/test/java/com/azure/messaging/eventhubs/implementation/instrumentation/EventHubsTracerTests.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import reactor.core.Exceptions;
 
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Stream;
@@ -30,8 +31,8 @@ public class EventHubsTracerTests {
         Tracer inner = mock(Tracer.class);
         when(inner.isEnabled()).thenReturn(true);
 
-        EventHubsTracer tracer = new EventHubsTracer(inner, "fqdn", "entityPath");
-        tracer.endSpan(null, Context.NONE, null);
+        EventHubsTracer tracer = new EventHubsTracer(inner, "fqdn", "entityPath", null);
+        tracer.endSpan(null, null, Context.NONE, null);
 
         verify(inner, times(1)).end(isNull(), isNull(), same(Context.NONE));
     }
@@ -41,9 +42,9 @@ public class EventHubsTracerTests {
         Tracer inner = mock(Tracer.class);
         when(inner.isEnabled()).thenReturn(true);
 
-        EventHubsTracer tracer = new EventHubsTracer(inner, "fqdn", "entityPath");
+        EventHubsTracer tracer = new EventHubsTracer(inner, "fqdn", "entityPath", null);
         AtomicBoolean closed = new AtomicBoolean();
-        tracer.endSpan(null, Context.NONE, () -> closed.set(true));
+        tracer.endSpan(null, null, Context.NONE, () -> closed.set(true));
 
         verify(inner, times(1)).end(isNull(), isNull(), same(Context.NONE));
         assertTrue(closed.get());
@@ -51,23 +52,29 @@ public class EventHubsTracerTests {
 
     @ParameterizedTest
     @MethodSource("getAmqpException")
-    public void testSpanEndException(Exception amqpException, String expectedStatus) {
+    public void testSpanEndException(Exception amqpException, Exception cause, String expectedStatus) {
         Tracer inner = mock(Tracer.class);
         when(inner.isEnabled()).thenReturn(true);
 
-        EventHubsTracer tracer = new EventHubsTracer(inner, "fqdn", "entityPath");
+        EventHubsTracer tracer = new EventHubsTracer(inner, "fqdn", "entityPath", null);
 
-        tracer.endSpan(amqpException, Context.NONE, null);
+        tracer.endSpan(null, amqpException, Context.NONE, null);
 
-        verify(inner, times(1)).end(eq(expectedStatus), eq(amqpException), same(Context.NONE));
+        verify(inner, times(1)).end(eq(expectedStatus), eq(cause), same(Context.NONE));
     }
 
     public static Stream<Arguments> getAmqpException() {
+        RuntimeException runtimeException = new RuntimeException("foo");
+        AmqpException amqpNoCauseNoCondition = new AmqpException(false, "foo", null, null);
+        AmqpException amqpNoCauseCondition = new AmqpException(false, AmqpErrorCondition.NOT_FOUND, "foo", null);
+        AmqpException amqpNoCauseConditionMessage = new AmqpException(false, AmqpErrorCondition.TIMEOUT_ERROR, "test", null);
+        AmqpException amqpCauseCondition = new AmqpException(false, AmqpErrorCondition.SERVER_BUSY_ERROR, null, runtimeException, null);
         return Stream.of(
-            Arguments.of(new RuntimeException("foo"), null),
-            Arguments.of(new AmqpException(false, "foo", null, null), null),
-            Arguments.of(new AmqpException(false, AmqpErrorCondition.NOT_FOUND, "foo", null), AmqpErrorCondition.NOT_FOUND.getErrorCondition()),
-            Arguments.of(new AmqpException(false, AmqpErrorCondition.TIMEOUT_ERROR, "", null), AmqpErrorCondition.TIMEOUT_ERROR.getErrorCondition()),
-            Arguments.of(new AmqpException(false, AmqpErrorCondition.SERVER_BUSY_ERROR, null, new RuntimeException("foo"), null), AmqpErrorCondition.SERVER_BUSY_ERROR.getErrorCondition()));
+            Arguments.of(runtimeException, runtimeException, RuntimeException.class.getName()),
+            Arguments.of(Exceptions.propagate(runtimeException), runtimeException, RuntimeException.class.getName()),
+            Arguments.of(amqpNoCauseNoCondition, amqpNoCauseNoCondition, AmqpException.class.getName()),
+            Arguments.of(amqpNoCauseCondition, amqpNoCauseCondition, amqpNoCauseCondition.getErrorCondition().getErrorCondition()),
+            Arguments.of(amqpNoCauseConditionMessage, amqpNoCauseConditionMessage, amqpNoCauseConditionMessage.getErrorCondition().getErrorCondition()),
+            Arguments.of(amqpCauseCondition, runtimeException, amqpCauseCondition.getErrorCondition().getErrorCondition()));
     }
 }